+++ lib/util.h
@@ -27,7 +27,8 @@ void split_string(const char* in, int *len, char***results);
 /* Takes a list of ports and port ranges as strings, and parses them
  * into a list of port_h_t.
  */
-void parse_ranges(char **in, int in_len, port_h_t **ports, int *ports_len);
+void parse_ranges(char **in, int in_len, port_h_t **ports,
+		uint16_t *ports_len);
 
 /* Takes a comma-seperated list of ports or port ranges and returns
  * a list of ports as port_h_t's. Intended to convert the command-line
